+++ src/saxhandler.cpp
@@ -366,7 +366,7 @@ void SAXSpectraHandler::decode32(bool bM /*= true*/, bool bI /*= true*/)
 		// Base64 decoding
 		// By comparing the size of the unpacked data and the expected size
 		// an additional check of the data file integrity can be performed
-		int length = b64_decode_mio( (char*) pDecoded , (char*) pData, stringSize );
+		size_t length = b64_decode_mio( (char*) pDecoded , (char*) pData, stringSize );
 		if(length != size) {
 //			cout << " decoded size " << length << " and required size " << (unsigned long)size << " dont match:\n";
 			Rprintf(" decoded size %i and required size %l dont match:\n", length, (unsigned long)size);
@@ -419,7 +419,7 @@ void SAXSpectraHandler::decode64(bool bM /*= true*/, bool bI /*= true*/)
 		// Base64 decoding
 		// By comparing the size of the unpacked data and the expected size
 		// an additional check of the data file integrity can be performed
-		int length = b64_decode_mio( (char*) pDecoded , (char*) pData, stringSize );
+		size_t length = b64_decode_mio( (char*) pDecoded , (char*) pData, stringSize );
 		if(length != size) {
 //			cout << " decoded size " << length << " and required size " << (unsigned long)size << " dont match:\n";
 			Rprintf(" decoded size %i and required size dont match:\n", length, (unsigned long)size);
